2023/04/10 이해빈 / res_no -> long으로 변경

+++ Lemo/src/main/java/kr/co/Lemo/controller/ProductController.java
@@ -188,7 +188,7 @@ public class ProductController {
             return "redirect:/product/list";
         }
 
-        int res_no = (int) obj;
+        long res_no = (long) obj;
         String user_id = myUser.getUser_id();
 
         // 로그인한 유저 id와 예약번호(res_no)로 조회하여 예약내역이 있는지 검증
